requirements.txt
fastapi~=0.104.0
uvicorn[standard]

python-dotenv~=1.0.0

langchain~=0.0.324
openai
tiktoken
chromadb
pypdf
